+++ apps/pixel-enrichment-function/src/apis/linkedin-url/proxycurl.person-lookup.ts
@@ -1,4 +1,5 @@
 import freeEmailDomains from 'free-email-domains';
+
 import type { PixelCollectionData } from '@kynesis/pixel-enrichment-sqs';
 
 import type { LinkedinUrlResponse } from './interfaces/response';