+++ apps/pixel-enrichment-function/src/apis/linkedin-url/rampedup.ts
@@ -1,4 +1,5 @@
 import freeEmailDomains from 'free-email-domains';
+
 import type { PixelCollectionData } from '@kynesis/pixel-enrichment-sqs';
 
 import type { LinkedinUrlResponse } from './interfaces/response';
@@ -11,7 +12,7 @@ class RampedupApi extends LinkedinUrl {
 	protected override apiUrl = 'https://basic.rampedup.io/api/manualappend/contact';
 
 	public override async getLinkedinUrl(data: PixelCollectionData): Promise<LinkedinUrlResponse> {
-		// * Email will be valid as it is being validated on "pixel-api-function"
+		// * Email will be valid as it is being validated on other Lambda functions
 		const emailDomain = data.email.split('@')[1]!;
 		const emailDomainSplitted = emailDomain.split('.');
 		const companyDomain = emailDomainSplitted.at(-2) + '.' + emailDomainSplitted.at(-1);